First draft of new way of running and outputting results.
We now use a queue rather then separating the MrBump jobs into batches
so that the work is fairly partitioned out.

On early_terminate, the other jobs are killed by ample rather then the entire script
being killed so we have a normal exit.

I'm also splitting the functions to check if a job has run and to print/output the table.