3 changes:
Changed last few options (allatom, frags3mers, frags9mers, molreponly, phaseronly, usehoms) so that they are now
underscore separated as are all other options

Changed how the options to ample are handled so that we now have a devel_mode and quick_mode with preset options.

Created a summary for the ensembles that goes into the ROSETTA.log file - also changed the information that ends up in this file.