Fixed bug/feature where we ignored subclusters that were the same size as previous subcluster when there are > ensemble_max_models in the subcluster.
We now remember which models are in any subcluster and only reject a subcluster if
the models in it are the same as any of the other subclusters.

For clusters with > ensemble_max_models we now randomly sample the subcluster to 
select the list of ensemble_max_models. This introduces an element of randomness
into ample runs that wasn't there previously.