Implemented logger as a singleton class as I now understand it better.
Each module should create its own logger with :

_logger = logging.getLogger(__name__)

This way, debug log messages from each module include the module name, which makes it a lot easier to spot where errors are happening.

Also split the creation of the log handlers, so that we create a stream logger at the start, so we therefore can eschew all print statements. Yay!