Enabled gesamt as the subclusterer by setting subcluster_program to gesamt
Moved find_maxcluster into maxcluster file to make it easier to retire if necessary.

Modified handling of nproc argument. Whole way we pass in or set arguments to the ensembler objects/methods
needs refactoring as it's a bit of a hybrid mess.

Removed some of the switch statements and replaced them with good old-fashioned if-else statements as otherwise
I get lost as to what's happening, and it's much harder to refactor and tweak things. Sorry Felix!