Refactored AMPLE ensembler code.
The new code makes greater use of objects to simplify the parameters passed between the 
different stages. There are now objects for Clusters and Ensembles, and a Truncator object
that creates Truncations.

Some docstring in numpy format have been added with the intention that we will eventually
have all major methods so documented to facilitate the creation of online documentation.