Created a MRinfo object that calculates the origin shift, weighted and normal mean phase error for an MR pdb.
The object can be set up with the data for native structure and then called simply for any MR pdb that uses the native mtz.

The benchmark code has been updated to use this and we now automatically capture the phase error for the MR pdb, SHELXE pdb and any BUCCANEER or ARPWARP pdb.